+++ Helper/Charge.php
@@ -152,7 +152,7 @@ class Charge extends AbstractHelper
             $log['response_errors'] = $this->client->getErrors();
             $this->logger->addError(__METHOD__, $log, true);
 
-            return false;
+            return $response;
         }
     }